+++ test/WebJobs.Script.Tests/Diagnostics/OpenTelemetry/OpenTelemetryConfigurationExtensionsTests.cs
@@ -5,6 +5,9 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
+using System.Reflection;
+using Azure.Identity;
+using Azure.Monitor.OpenTelemetry.Exporter;
 using FluentAssertions;
 using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.Extensibility;
@@ -231,6 +234,78 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Diagnostics.OpenTelemetry
             Assert.Equal(4, resource.Attributes.Count());
         }
 
+        [Fact]
+        public void ConfigureTelemetry_Should_UseOpenTelemetryWhenModeSetAndAppInsightsAuthStringClientIdPresent()
+        {
+            // Arrange
+            var clientId = Guid.NewGuid();
+            IServiceCollection serviceCollection = default;
+
+            var hostBuilder = new HostBuilder()
+                .ConfigureAppConfiguration(config =>
+                {
+                    config.AddInMemoryCollection(new Dictionary<string, string>
+                    {
+                        { "APPLICATIONINSIGHTS_AUTHENTICATION_STRING", $"Authorization=AAD;ClientId={clientId}" },
+                        { "APPLICATIONINSIGHTS_CONNECTION_STRING", "InstrumentationKey=key" },
+                        { ConfigurationPath.Combine(ConfigurationSectionNames.JobHost, "telemetryMode"), TelemetryMode.OpenTelemetry.ToString() }
+                    });
+                })
+                .ConfigureDefaultTestWebScriptHost()
+                .ConfigureLogging((context, loggingBuilder) => loggingBuilder.ConfigureTelemetry(context))
+                .ConfigureServices(services => serviceCollection = services);
+
+            using var host = hostBuilder.Build();
+
+            // Act
+            var tracerProviderDescriptors = GetTracerProviderDescriptors(serviceCollection);
+            var resolvedClient = ExtractClientFromDescriptors(tracerProviderDescriptors);
+
+            // Extract the clientId from the client object
+            var clientIdValue = resolvedClient?.GetType().GetProperty("ClientId", BindingFlags.Instance | BindingFlags.NonPublic)?.GetValue(resolvedClient)?.ToString();
+
+            // Assert
+            serviceCollection.Should().NotBeNullOrEmpty();
+            clientIdValue.Should().Be(clientId.ToString());
+            resolvedClient.GetType().Name.Should().Be("ManagedIdentityClient");
+        }
+
+        [Fact]
+        public void ConfigureTelemetry_Should_UseOpenTelemetryWhenModeSetAndAppInsightsAuthStringPresent()
+        {
+            // Arrange
+            IServiceCollection serviceCollection = default;
+
+            var hostBuilder = new HostBuilder()
+                .ConfigureAppConfiguration(config =>
+                {
+                    config.AddInMemoryCollection(new Dictionary<string, string>
+                    {
+                        { "APPLICATIONINSIGHTS_AUTHENTICATION_STRING", $"Authorization=AAD" },
+                        { "APPLICATIONINSIGHTS_CONNECTION_STRING", "InstrumentationKey=key" },
+                        { ConfigurationPath.Combine(ConfigurationSectionNames.JobHost, "telemetryMode"), TelemetryMode.OpenTelemetry.ToString() }
+                    });
+                })
+                .ConfigureDefaultTestWebScriptHost()
+                .ConfigureLogging((context, loggingBuilder) => loggingBuilder.ConfigureTelemetry(context))
+                .ConfigureServices(services => serviceCollection = services);
+
+            using var host = hostBuilder.Build();
+
+            // Act
+            var tracerProviderDescriptors = GetTracerProviderDescriptors(serviceCollection);
+            var resolvedClient = ExtractClientFromDescriptors(tracerProviderDescriptors);
+
+            // Extract the clientId from the client object
+            var clientIdValue = resolvedClient?.GetType().GetProperty("ClientId", BindingFlags.Instance | BindingFlags.NonPublic)?.GetValue(resolvedClient)?.ToString();
+
+            // Assert
+            serviceCollection.Should().NotBeNullOrEmpty();
+            // No clientId should be present as it was not provided
+            clientIdValue.Should().BeNull();
+            resolvedClient.GetType().Name.Should().Be("ManagedIdentityClient");
+        }
+
         // The OpenTelemetryEventListener is fine because it's a no-op if there are no otel events to listen to
         private bool HasOtelServices(IServiceCollection sc) => sc.Any(sd => sd.ServiceType != typeof(OpenTelemetryEventListener) && sd.ServiceType.FullName.Contains("OpenTelemetry"));
 
@@ -243,6 +318,47 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Diagnostics.OpenTelemetry
             { "WEBSITE_OWNER_NAME", "AAAAA-AAAAA-AAAAA-AAA+appName-EastUSwebspace" },
             { "REGION_NAME", "EastUS" }
         });
+        }
+
+        private static List<ServiceDescriptor> GetTracerProviderDescriptors(IServiceCollection services)
+        {
+            return services
+                .Where(descriptor =>
+                    descriptor.Lifetime == ServiceLifetime.Singleton &&
+                    descriptor.ServiceType.Name == "IConfigureTracerProviderBuilder" &&
+                    descriptor.ImplementationInstance?.GetType().Name == "ConfigureTracerProviderBuilderCallbackWrapper")
+                .ToList();
+        }
+
+        private static object ExtractClientFromDescriptors(List<ServiceDescriptor> descriptors)
+        {
+            foreach (var descriptor in descriptors)
+            {
+                var implementation = descriptor.ImplementationInstance;
+                if (implementation is null)
+                {
+                    continue;
+                }
+
+                // Reflection starts here
+                var configureField = implementation.GetType().GetField("configure", BindingFlags.Instance | BindingFlags.NonPublic);
+                if (configureField?.GetValue(implementation) is Action<IServiceProvider, TracerProviderBuilder> configureDelegate)
+                {
+                    var targetType = configureDelegate.Target.GetType();
+                    var configureDelegateTarget = targetType.GetField("configure", BindingFlags.Instance | BindingFlags.Public);
+
+                    if (configureDelegateTarget?.GetValue(configureDelegate.Target) is Action<AzureMonitorExporterOptions> exporterOptionsDelegate)
+                    {
+                        var credentialField = exporterOptionsDelegate.Target.GetType().GetField("credential", BindingFlags.Instance | BindingFlags.Public);
+                        if (credentialField?.GetValue(exporterOptionsDelegate.Target) is ManagedIdentityCredential managedIdentityCredential)
+                        {
+                            var clientProperty = managedIdentityCredential.GetType().GetProperty("Client", BindingFlags.Instance | BindingFlags.NonPublic);
+                            return clientProperty?.GetValue(managedIdentityCredential);
+                        }
+                    }
+                }
+            }
+            return null;
         }
     }
 }