+++ release_notes.md
@@ -3,7 +3,8 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
-
+- Add support for managed identity when using open telemetry + azure monitor (#10615)
+- Update Java Worker Version to [2.18.0](https://github.com/Azure/azure-functions-java-worker/releases/tag/2.18.0)
 - Allow for an output binding value of an invocation result to be null (#10698)
 - Updated dotnet-isolated worker to 1.0.12.
   - [Corrected the path for the prelaunch app location.](https://github.com/Azure/azure-functions-dotnet-worker/pull/2897)